requirements.txt
customtkinter
pandas
gspread
pyodbc
python-dotenv
oauth2client
